Fix persistence MongoDB

+++ app.js
@@ -1,5 +1,7 @@
 require('dotenv').config();
 const Discord = require('discord.js');
+const blacklist = require('./src/db/blacklist');
+const whitelist = require('./src/db/whitelist');
 const bot = new Discord.Client();
 
 const message = require('./src/message');
@@ -15,10 +17,20 @@ bot.on('ready', () => {
 });
 
 bot.on('message', msg => {
-    if (cleanChannels.indexOf(msg.channel.id) !== -1) {
+    if (cleanChannels.indexOf(msg.channel.id) !== -1 && !msg.author.bot) {
 
-        message.checkMessage(msg.content).then(res => {
-            msg.react('✅');
+        message.checkMessage(msg.content).then(url => {
+            if (msg.channel.id === process.env.WHITELIST_CHANNEL_ID) {
+                whitelist.add(url).then(() => {
+                    msg.react('✅');
+                }).catch(msg.author.send);
+            } else if (msg.channel.id === process.env.BLACKLIST_CHANNEL_ID) {
+                blacklist.add(url).then(() => {
+                    msg.react('✅');
+                }).catch(msg.author.send);
+            } else if (msg.channel.id === process.env.TEST_CHANNEL_ID) {
+                msg.react('💩');
+            }
         }).catch(error => {
             msg.author.send(error.message);
             msg.delete();
